Added Via Upload
Added a new version of the flexdata separator smooth that writes driver functions into a ready-to-run file in Blender if expfix is set to 1. That way, this file can be linked into a new scene when working with industry workflows. This file can be registered, so it will apply all the needed functions every time Blender opens the scene. This file also contains my fix for when slider definitions exceed 63 characters, which are hard-coded into Blender, renames them, and uses them when applying the drivers, so this tool can be used with custom and other Valve HWM models and alike that use longer expression names.